Create Linked List Cycle.java
Linked List Cycle

You are given the head node of the linked list. You need to check for any cycle is present in linked list or not, You need to output 1 if cycle is present else 0.

Note:
You need to only complete the function. Dont worry about input, it is for internal reference.

Input Format
The first line of input contains N representing the number of nodes in linked list.

The second line of input contains N space separated integers, representing elements in linked list.

The third line of input contains a number X representing which node is connected to others.

Output Format
In a single line print 1 or 0.

Example 1
Input

3
1 3 4
2
Output

1
Explanation

1->3->4
  ^   |
  |___|
This is the list given in question. A loop is present in this linked list

Example 2
Input

4
1 2 3 4
0
Output

0
Explanation

1->2->3->4 this is list you are given in question, and the answer is 0 as it doesnt contains any cycle.

Constraints
1 <= N <= 1000

1 <= value of node <= 1000