Create Merge two sorted linked list.java
Merge two sorted linked list
Merge two sorted linked lists and return it as a sorted list. The list should be made by splicing together the nodes of the first two lists.

Input Format
The format for each test case is as follows:

The first line contains an integer n, the length of the first linked list.
The next line contain n integers, the elements of the linked list.
The next line contains an integer m , the length of the second linked list.
The next lines contain m integers, the elements of the second linked list.

Output Format
Output a single line of (n + m) integers consisting all elements of linked lists in sorted order.

Example 1
Input
3
1 2 4
3
1 3 4

Output
1 1 2 3 4 4
Merge the two linked list

Example 2
Input
3
1 5 9
3
1 3 4

Output
1 1 3 4 5 9
Explanation

Merge the two linked list in sorted order

Constraints
The number of nodes in both lists is in the range [0, 50].

-100 <= Node.val <= 100

Both list1 and list2 are sorted in non-decreasing order.