Create Intersection of Two Linked Lists.java
Intersection of Two Linked Lists
Your are given two linked List consisting of N nodes and M nodes consisting of a common intersection point.

Your task is to find the value of the intersection point of the two linked list.

Input Format
The first line contains the number of test cases.

The first line of each test case has an integer containing the number of elements in first linked list , the number of elements in second linked list and a hidden integer k that is a random index of first linked list.

Next line denotes the elements of the linked list.

Next line denotes the elements of the next linked list.

All the elements of first list beyond the node with index k are appended to second list.

Output Format
For each test case print an integer in a new line, denoting the value of the intersection point.

Example 1
Input

1
5 1 3
3 6 9 15 30
10
Output

15
Explanation

L1: 3 - 6 - 9 - 15 - 30 

L2: 10 - 15 - 30

3
 \
  6
   \
    9    10   
     \  /
      15
      |
      30

The value of the intersection node is 15. Thus answer is 15

Example 2
Input

1
5 1 3
1 2 3 4 5
3
Output

4
Explanation

The value of intersection node is 4. Thus answer is 4.

Constraints
1 <= T <= 10

1 <= N <= 10000

1 <= M <= 10000

1 <= L1[i] <= 100000

1 <= L2[i] <= 100000