Create Find length of Loop.java
Find length of Loop
Given a linked list of size N. The task is to complete the function countNodesinLoop() that checks whether a given Linked List contains a loop or not and if the loop is present then return the count of nodes in a loop or else return 0. C is the position of the node to which the last node is connected. If it is 0 then no loop.

Input Format
The first line contains a single integer N which denotes the size of the linked list. Next line contains N spaced integers which represent the elements of the linked list. Next line contains a single integer C which denotes the position of the node to which the last node is connected.

Output Format
Print the length of the loop (0 if there is no loop).

Example 1
Input

4
1 2 3 4
2
Output

3
Explanation

1-2-3-4
  |___|
The length of the loop is 3. 
Example 2
Input

5
1 2 3 4 5
0
Output

0
Explanation

Since no loop present in the linked list. Hence, the answer is 0.
Constraints
1 <= N <= 500

0 <= C <= N-1