Create Flattening a Linked List.java
Flattening a Linked List
Given a Linked List of size N, where every node represents a sub-linked-list and contains two pointers:

Pointer to next node in the main list (we call it ‘right’ pointer in the code below)
Pointer to a linked list where this node is headed (we call it the ‘down’ pointer in the code below).
Each of the sub-linked-list is in sorted order. Flatten the Link List such that all the nodes appear in a single level while maintaining the sorted order.

Note: The flattened list will be printed using the down pointer instead of the right pointer.

For more clarity have a look at the printList() function in the driver code.

Input Format
First line consits of an integer n, the number of groups of chains in the linked list.

For the next n lines, each input consists of an integer k, followed by k nodes.

Output Format
Just flatten the linked list and print it and the compiler will print 'yes' if your output is correct else it prints 'no'

Example 1
Input


4
4
5 7 8 30
2
10 20
3
19 22 50
4
28 35 40 45
Output


5 7 8 10 19 20 22 28 30 35 40 45 50
yes
Explanation


The resultant linked list has every node in a single level. 

Given LinkedList : 

5-->10-->19-->28
|    |    |    |
7   20   22   35
|         |    |
8        50   40
|              |
30            45

Sorted LinkedList  : 5 7 8 10 19 20 22 28 30 35 40 45 50  
Example 2
Input


4
4
5 7 8 30
1
10
3
19 22 50
1
28
Output


5 7 8 10 19 22 28 30 50
yes 
Explanation


The resultant linked list has every node in a single level.

Given LinkedList : 

5-->10-->19-->28
|         |    
7        22   
|         |    
8        50   
|              
30            

Sorted LinkedList  : 5 7 8 10 19 22 28 30 50 
Constraints
0 <= n <= 50

1 <= Ki <= 20

1 <= Element of linked list <= 10^3